Implement DNS record verification
Checks to make sure DNS indicates an A/AAAA record matching the
public IP address of the server before allowing the domain name
to be added.

Uses a random host to verify wildcard domains.

Might be more robust to actually communicate with the
authoritative servers, or check from multiple geo locations, but
this is probably fine for now and maybe forever.